data_scrapping/requirement.txt
beautifulsoup4
requests
pymysql
urllib3
pandas
urllib3==1.26.6
psycopg2-binary
unidecode